backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Cache
pymongo==4.6.0  # Includes bson module
redis==5.0.1

# AI & ML
google-generativeai>=0.3.0  # Google Gemini SDK
sentence-transformers>=2.3.0,<3.0.0  # Compatible with modern huggingface-hub
huggingface-hub>=0.20.0  # Required for sentence-transformers 2.3+
transformers>=4.21.0  # Required by sentence-transformers
numpy==1.24.3
torch>=1.13.0  # Required by sentence-transformers

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
